backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
psycopg2-binary
python-dotenv
drf-yasg
datetime
humanize